+++ TinyRenderer/triangle.h
@@ -4,4 +4,4 @@
 
 Vec3f barycentric(Vec3f p2, Vec3f p1, Vec3f p0, Vec3f point);
 
-void rasterize(Triangle triangle, BMPImage& image, float* zBuffer, std::vector<FragmentShader*> shaders);
+void rasterize(Triangle triangle, BMPImage& image, int* zBuffer, std::vector<FragmentShader*> shaders);
